# PhaseIV/specs/001-todo-web-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Web App API
  description: API for managing user tasks in the Todo Web App
  version: 1.0.0
  contact:
    name: Todo Web App Team
servers:
  - url: https://api.todowebapp.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
  /auth/signup:
    post:
      summary: Register a new user
      description: Creates a new user account with secure password hashing
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticates user credentials and returns JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for authentication
  schemas:
    UserRegistration:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password (will be hashed)
        first_name:
          type: string
          maxLength: 50
          description: User's first name (optional)
        last_name:
          type: string
          maxLength: 50
          description: User's last name (optional)
      required:
        - email
        - password
    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
      required:
        - email
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User's unique identifier
        email:
          type: string
          format: email
          description: User's email address
        first_name:
          type: string
          description: User's first name
        last_name:
          type: string
          description: User's last name
        created_at:
          type: string
          format: date-time
          description: When the user account was created
        updated_at:
          type: string
          format: date-time
          description: When the user account was last updated
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          enum: [bearer]
          description: Type of token
        user:
          $ref: '#/components/schemas/UserResponse'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details